discord-feedback-bot/requirements.txt
discord.py>=2.6,<3.0
python-dotenv>=1.0,<2.0
Pillow>=11.0,<12.0
